+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep A Changelog's Format](http://keepachangelog.com/).
 
+## [1.3.1] 2017-07-28
+### Added
+- `ensureType` optional parameter to `sanitizeFileName`
+
 ## [1.3.0] 2017-02-15 
-## Added
+### Added
 - `sanitizeFileName.js` for string normalizing
 
 ## [1.2.0] 2017-01-10